requirements.txt
speechrecognition
pyttsx3
wikipedia
pywhatkit
pyaudio
requests
python-dotenv
pyjokes
cohere
python-dotenv


